Handle absolute, asterisk, and authority-form request targets
Improves compliance with RFC 7230 on the expected handling of requests that have URI or asterisk in the request target.

This means rejecting asterisk requests that are not OPTIONS and rejecting authority-form requests taht are not CONNECT.

This also means the server will handle the path and query on targets with absolute URIs as request-targets.